+++ compressor.cpp
@@ -4,34 +4,79 @@
 compressor::compressor(sector_tools_compression mode, bool is_compression, int8_t comp_level) {
     comp_mode = mode;
     compression = is_compression;
+    int ret;
     // Class initialzer
-    if (mode == C_ZLIB) {
-        strm.zalloc = Z_NULL;
-        strm.zfree = Z_NULL;
-        strm.opaque = Z_NULL;
+    switch(mode) {
+    case C_ZLIB:
+        strm_zlib.zalloc = Z_NULL;
+        strm_zlib.zfree = Z_NULL;
+        strm_zlib.opaque = Z_NULL;
 
-        int ret;
         if (is_compression) {
-            ret = deflateInit(&strm, comp_level);
+            ret = deflateInit(&strm_zlib, comp_level);
         }
         else {
-            ret = inflateInit(&strm);
+            ret = inflateInit(&strm_zlib);
         }
         
         if (ret != Z_OK) {
             //throw std::runtime_error("Error initializing the zlib compressor/decompressor.");
         }
+
+        break;
+
+    case C_LZMA:
+        strm_lzma = LZMA_STREAM_INIT;
+
+        if (is_compression) {
+            lzma_lzma_preset(&opt_lzma2, 9);
+
+            lzma_filter filters[] = {
+                { LZMA_FILTER_X86, NULL },
+                { LZMA_FILTER_LZMA2, &opt_lzma2 },
+                { LZMA_VLI_UNKNOWN, NULL },
+            };
+            ret = lzma_stream_encoder(&strm_lzma, filters, LZMA_CHECK_NONE); // CRC is already checked
+
+            // Return successfully if the initialization went fine.
+            if (ret != LZMA_OK) {
+                //throw std::runtime_error("Error initializing the lzma2 compressor/decompressor.");
+            }
+        }
+        else {
+            ret = lzma_stream_decoder(
+                &strm_lzma,
+                UINT64_MAX,
+                LZMA_IGNORE_CHECK
+            );
+
+            // Return successfully if the initialization went fine.
+            if (ret != LZMA_OK) {
+                //throw std::runtime_error("Error initializing the lzma2 compressor/decompressor.");
+            }
+        }
+
+        break;
     }
 }
 
 
 int8_t compressor::set_input(uint8_t* in, size_t &in_size){
     if (!compression) {
-        if (comp_mode == C_ZLIB) {
-            strm.avail_in = in_size;
-            strm.next_in = in;
+        switch(comp_mode) {
+        case C_ZLIB:
+            strm_zlib.avail_in = in_size;
+            strm_zlib.next_in = in;
 
             return 0;
+            break;
+
+        case C_LZMA:
+            strm_lzma.avail_in = in_size;
+            strm_lzma.next_in = in;
+
+            return 0;
+            break;
         }
 
         return 0;
@@ -45,11 +90,20 @@ int8_t compressor::set_input(uint8_t* in, size_t &in_size){
 
 int8_t compressor::set_output(uint8_t* out, size_t &out_size){
     if (compression) {
-        if (comp_mode == C_ZLIB) {
-            strm.avail_out = out_size;
-            strm.next_out = out;
+        switch(comp_mode) {
+        case C_ZLIB:
+            strm_zlib.avail_out = out_size;
+            strm_zlib.next_out = out;
 
             return 0;
+            break;
+
+        case C_LZMA:
+            strm_lzma.avail_out = out_size;
+            strm_lzma.next_out = out;
+
+            return 0;
+            break;
         }
 
         return 0;
@@ -61,16 +115,41 @@ int8_t compressor::set_output(uint8_t* out, size_t &out_size){
 }
 
 
-int8_t compressor::compress(size_t &out_size, uint8_t* in, size_t in_size, uint8_t flusmode){
+int8_t compressor::compress(size_t &out_size, uint8_t* in, size_t in_size, uint8_t flushmode){
     if (compression) {
-        if (comp_mode == C_ZLIB) {
-            strm.avail_in = in_size;
-            strm.next_in = in;
+        int8_t return_code;
 
-            int8_t return_code = deflate(&strm, flusmode);
+        switch(comp_mode) {
+        case C_ZLIB:
+            strm_zlib.avail_in = in_size;
+            strm_zlib.next_in = in;
 
-            out_size = strm.avail_out;
+            return_code = deflate(&strm_zlib, flushmode);
+
+            out_size = strm_zlib.avail_out;
             return return_code;
+            break;
+
+        case C_LZMA:
+            strm_lzma.avail_in = in_size;
+            strm_lzma.next_in = in;
+
+            lzma_action flushmode_lzma = LZMA_RUN;
+            switch (flushmode) {
+                case Z_FULL_FLUSH:
+                    flushmode_lzma = LZMA_FULL_FLUSH;
+                    break;
+
+                case Z_FINISH:
+                    flushmode_lzma = LZMA_FINISH;
+                    break;
+            }
+
+            return_code = lzma_code(&strm_lzma, flushmode_lzma);
+
+            out_size = strm_lzma.avail_out;
+            return return_code;
+            break;
         }
 
         return 0;
@@ -83,13 +162,25 @@ int8_t compressor::compress(size_t &out_size, uint8_t* in, size_t in_size, uint8
 
 int8_t compressor::decompress(uint8_t* out, size_t out_size, size_t &in_size, uint8_t flusmode){
     if (!compression) {
-        if (comp_mode == C_ZLIB) {
-            strm.avail_out = out_size;
-            strm.next_out = out;
-            int8_t return_code = inflate(&strm, flusmode);
+        int8_t return_code;
+        switch(comp_mode) {
+        case C_ZLIB:
+            strm_zlib.avail_out = out_size;
+            strm_zlib.next_out = out;
+            return_code = inflate(&strm_zlib, flusmode);
 
-            in_size = strm.avail_in;
+            in_size = strm_zlib.avail_in;
             return return_code;
+            break;
+
+        case C_LZMA:
+            strm_lzma.avail_out = out_size;
+            strm_lzma.next_out = out;
+            return_code = lzma_code(&strm_lzma, LZMA_RUN);
+
+            in_size = strm_lzma.avail_in;
+            return return_code;
+            break;
         }
 
         return 0;
@@ -101,13 +192,53 @@ int8_t compressor::decompress(uint8_t* out, size_t out_size, size_t &in_size, ui
 }
 
 int8_t compressor::close(){
-    if (compression) {
-        (void)deflateEnd(&strm);
-        strm = {};
+    switch(comp_mode) {
+    case C_ZLIB:
+        if (compression) {
+            (void)deflateEnd(&strm_zlib);
+            strm_zlib = {};
+        }
+        else {
+            (void)inflateEnd(&strm_zlib);
+            strm_zlib = {};
+        }
+        return 0;
+        break;
+
+    case C_LZMA:
+        lzma_end(&strm_lzma);
+        strm_lzma = {};
+        return 0;
+        break;
     }
-    else {
-        (void)inflateEnd(&strm);
-        strm = {};
-    }
-    return 0;
+
+    return -1;
 }
+
+size_t compressor::data_left_in() {
+    switch(comp_mode) {
+    case C_ZLIB:
+        return strm_zlib.avail_in;
+        break;
+
+    case C_LZMA:
+        return strm_lzma.avail_in;
+        break;
+    }
+
+    return -1;
+};
+
+size_t compressor::data_left_out() {
+    switch(comp_mode) {
+    case C_ZLIB:
+        return strm_zlib.avail_out;
+        break;
+
+    case C_LZMA:
+        return strm_lzma.avail_out;
+        break;
+    }
+
+    return -1;
+};